requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
deepagents==0.1.0
requests==2.31.0
langchain-openai>=0.2.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
streamlit==1.39.0
